Update behavioral-interview-questions.md
Airbnb is a horrible company with shady business practices that are unfavorable to the common traveller.  This pull request is to tweak this otherwise fantastic document to put fewer eyes on the lodging broker.  Anyone with a grain of ethics and integrity wouldn't want to work there, anyway, as it would be huge conflict of interest.